Update minSDK to 14
Android Studio was giving a warning that 8 was too low and so I kept incrementing until it stopped complaining at 14.  I wanted to test this but there is no emulator for 14 and the emulator for 15 didn't work for me.  I was able to get emulator for 16 to work and so I'm setting the min to that.