Show messages regarding network connectivity
If no network, show a toast and immediately stop trying to connect

If connected, but not to wifi, show a warning but keep trying. There's nothing preventing the player from working on a mobile data network, but I don't think that's a common use case. So it's probably more help to warn about wifi.

If connected to wifi, continue as normal.