Attempt to fix issues with edge to edge
Supporting edge to edge has proven to be difficult. Issues with light mode and the themes and the status bar result in a poor looking UI. The status bar seems to default to using white icons and I'm unable to figure out how to make them darker.

The easiest fix found so far is to force dark mode.